Use fixed type int64 length and positions

+++ jfile.h
@@ -10,28 +10,28 @@ struct JFile {
   std::FILE* jf = nullptr;
 
   // Current position of the main file.
-  long pos = 0;
+  int64_t pos = 0;
 
   // Max position of the main file that we can seek to.
-  long maxPos = 0;
+  int64_t maxPos = 0;
 
-  long lastPersistedPos = 0;
-  long lastPersistedMaxPos = 0;
+  int64_t lastPersistedPos = 0;
+  int64_t lastPersistedMaxPos = 0;
 
   // The ending position of the journal.
   // The actual file might be larger than this,
   // since we don't clear the journal after flushing.
-  long journalEndPos = 0;
+  int64_t journalEndPos = 0;
 
   // Number of completed blocks
-  int64_t numCompletedBlocks = 0;
+  size_t numCompletedBlocks = 0;
 
   // Starting position of the current block
   // in the journal file
-  long journalBlockStartPos = 0;
+  int64_t journalBlockStartPos = 0;
 
   // Length of the current block, start counting
   // from journal block start (i.e. including block header bytes).
-  long currentBlockLength = 0;
+  int64_t currentBlockLength = 0;
 };
 }
